Abandon using the tail command
The tail command didn't work right on Cygwin, so just implement our own simple
and stupid tail command.

In addition, I rearranged the code a little to proceed in a more logical
order, and accounted for a difference between Cygwin and Linux wrt the
implementation of the fpos_t type.  We are peaking into what should be
an opaque data structure.  Shame on us.