Move the stop_search_on_match attribute to the logfile where it can move around
This changes where the search order is defined.  I had originally
intended the search order to be determined by the order of audio triggers
in the atconfig.xml file, but an optimization in the implementation made me
make it so the attached triggers to a log file could be in an arbitrary
order, and that order determined the search order.

This turns out to provide more flexibility for the user, and allows
them to have a different search order and "stop search" settings on
different log files.